requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
python-dotenv==0.17.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
spotipy==2.18.0
urllib3==1.26.5
